Convert despacer to a normal library
- This makes it possible to make a static library of the despacer
- This allows linking despacer with other languages such as D, Rust, etc
