--- src/lib.c
@@ -1,1 +1,0 @@
-#include "../include/despacer.h"
